Adding trace-level log info to HTTP spans
