+++ apm-agent-core/src/main/java/co/elastic/apm/agent/impl/transaction/TraceContext.java
@@ -539,6 +539,9 @@ public class TraceContext implements Recyclable {
             String tracestateHeaderValue = TextTracestateAppender.instance().join(tracestate, coreConfiguration.getTracestateSizeLimit());
             headerSetter.setHeader(TRACESTATE_HEADER_NAME, tracestateHeaderValue, carrier);
         }
+        if (logger.isTraceEnabled()) {
+            logger.trace("Trace context headers added to {}", carrier);
+        }
     }
 
     /**
