+++ apm-agent-plugins/apm-spring-resttemplate-plugin/src/main/java/co/elastic/apm/agent/resttemplate/SpringRestTemplateInstrumentation.java
@@ -25,21 +25,12 @@
 package co.elastic.apm.agent.resttemplate;
 
 import co.elastic.apm.agent.bci.TracerAwareInstrumentation;
-import co.elastic.apm.agent.http.client.HttpClientHelper;
-import co.elastic.apm.agent.impl.transaction.AbstractSpan;
-import co.elastic.apm.agent.impl.transaction.Span;
-import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-import org.springframework.http.client.ClientHttpRequest;
-import org.springframework.http.client.ClientHttpResponse;
 
-import javax.annotation.Nullable;
-import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.Objects;
 
 import static net.bytebuddy.matcher.ElementMatchers.declaresMethod;
 import static net.bytebuddy.matcher.ElementMatchers.hasSuperType;
@@ -76,41 +67,5 @@ public class SpringRestTemplateInstrumentation extends TracerAwareInstrumentatio
     @Override
     public Collection<String> getInstrumentationGroupNames() {
         return Arrays.asList("http-client", "spring-resttemplate");
-    }
-
-    public static class SpringRestTemplateAdvice {
-        @Nullable
-        @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
-        public static Object beforeExecute(@Advice.This ClientHttpRequest request) {
-            if (tracer.getActive() == null) {
-                return null;
-            }
-            final AbstractSpan<?> parent = tracer.getActive();
-            Span span = HttpClientHelper.startHttpClientSpan(parent, Objects.toString(request.getMethod()), request.getURI(), request.getURI().getHost());
-            if (span != null) {
-                span.activate();
-                span.propagateTraceContext(request, SpringRestRequestHeaderSetter.INSTANCE);
-                return span;
-            }
-            return null;
-        }
-
-        @Advice.OnMethodExit(suppress = Throwable.class, onThrowable = Throwable.class, inline = false)
-        public static void afterExecute(@Advice.Return @Nullable ClientHttpResponse clientHttpResponse,
-                                        @Advice.Enter @Nullable Object spanObj,
-                                        @Advice.Thrown @Nullable Throwable t) throws IOException {
-            if (spanObj instanceof Span) {
-                Span span = (Span) spanObj;
-                try {
-                    if (clientHttpResponse != null) {
-                        int statusCode = clientHttpResponse.getRawStatusCode();
-                        span.getContext().getHttp().withStatusCode(statusCode);
-                    }
-                    span.captureException(t);
-                } finally {
-                    span.deactivate().end();
-                }
-            }
-        }
     }
 }
